move log markers in LogUtils.java and move DataFormatUtils.java to utils package
